refactor[#40] : entities/article post 요청 index.ts 파일로 이동
