+++ blog/src/entities/article/model/index.ts
@@ -1,176 +1,17 @@
-import { ARTICLE_ENDPOINT, ITEM_PER_PAGE } from "../config";
+import { ITEM_PER_PAGE } from "../config";
 import { ARTICLE_QUERY_KEY, type ArticleStatus } from "./articleQueryKey";
 import {
-  useMutation,
-  useQueryClient,
   useSuspenseInfiniteQuery,
   useSuspenseQuery
 } from "@tanstack/react-query";
-
-import { compressImage } from "@/entities/image/lib";
-import { Tag } from "@/entities/tag/@x/article";
 
 import { createBrowserSupabase } from "@/shared/model";
 import { snakeToCamel } from "@/shared/util";
 
 export * from "./getPolularArticleList";
-
-export interface PostArticleImageRequest {
-  files: File[];
-  articleId: string;
-}
-
-export interface PostArticleImageResponse {
-  code: number;
-  message: string;
-  data: {
-    path: string;
-    fullPath: string;
-    id: string;
-    imageUrl: string;
-  }[];
-}
-
-// TODO useMarkdown 훅 리팩토링 할 때 수정하기
-export const postArticleImage = async ({
-  files,
-  articleId
-}: PostArticleImageRequest) => {
-  const form = new FormData();
-
-  form.append("articleId", articleId);
-
-  files.forEach((file) => {
-    form.append("image", file);
-  });
-
-  const response = await fetch(ARTICLE_ENDPOINT.POST_ARTICLE_IMAGE(), {
-    method: "POST",
-    body: form
-  });
-
-  const { data, message } = (await response.json()) as PostArticleImageResponse;
-
-  if (!response.ok) {
-    throw new Error(message);
-  }
-
-  return data;
-};
-
-export interface PostArticleThumbnailRequest {
-  file: File;
-  articleId: string;
-}
-
-export interface PostArticleThumbnailResponse {
-  code: number;
-  message: string;
-  data: {
-    path: string;
-    fullPath: string;
-    id: string;
-    imageUrl: string;
-  };
-}
-
-const postArticleThumbnail = () => {
-  const mutationFn = async ({
-    file,
-    articleId
-  }: PostArticleThumbnailRequest) => {
-    const compressedImage = await compressImage(file, {
-      quantity: 2 ** 15
-    });
-
-    const formData = new FormData();
-    formData.append("image", compressedImage);
-    formData.append("articleId", articleId);
-
-    const response = await fetch(ARTICLE_ENDPOINT.POST_ARTICLE_THUMBNAIL(), {
-      method: "POST",
-      body: formData
-    });
-
-    const { data, message } =
-      (await response.json()) as PostArticleThumbnailResponse;
-
-    if (!response.ok) {
-      throw new Error(message);
-    }
-
-    return data;
-  };
-
-  return { mutationFn };
-};
-
-export const usePostArticleThumbnail = () => {
-  return useMutation(postArticleThumbnail());
-};
-
-/**
- * created_at , updated_at은 서버에서 자동으로 생성
- */
-export interface PostNewArticleRequest {
-  // articleId
-  id: number;
-  title: string;
-  content: string;
-  author: string;
-  seriesName: string;
-  description: string;
-  tags: Tag[];
-  status: ArticleStatus;
-  thumbnailUrl: string | null;
-}
-
-export interface PostNewArticleResponse {
-  status: number;
-  message: string;
-  data: {
-    type: PostNewArticleRequest["status"];
-  };
-}
-
-const postNewArticle = () => {
-  const mutationFn = async (body: PostNewArticleRequest) => {
-    const response = await fetch(ARTICLE_ENDPOINT.POST_NEW_ARTICLE(), {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json"
-      },
-      body: JSON.stringify(body)
-    });
-
-    const { status, message, data } =
-      (await response.json()) as PostNewArticleResponse;
-
-    if (status > 200) {
-      throw new Error(message);
-    }
-    return {
-      status,
-      message,
-      data
-    };
-  };
-
-  return { mutationFn };
-};
-
-export const usePostNewArticle = () => {
-  const queryClient = useQueryClient();
-  return useMutation({
-    ...postNewArticle(),
-    onSuccess: ({ data: { type } }) => {
-      const queryKey = ARTICLE_QUERY_KEY.default(type);
-      queryClient.invalidateQueries({
-        queryKey
-      });
-    }
-  });
-};
+export * from "./postArticleImage";
+export * from "./postArticleThumbnail";
+export * from "./postNewArticle";
 
 export const getArticleList = (status: ArticleStatus) => {
   const queryKey = ARTICLE_QUERY_KEY.list_all(status);
